Group language configuration types together
