+++ src/vs/editor/common/modes.ts
@@ -661,104 +661,6 @@ export interface CodeLensProvider {
 	resolveCodeLens?(model: editorCommon.IReadOnlyModel, codeLens: ICodeLensSymbol, token: CancellationToken): ICodeLensSymbol | Thenable<ICodeLensSymbol>;
 }
 
-/**
- * A tuple of two characters, like a pair of
- * opening and closing brackets.
- */
-export type CharacterPair = [string, string];
-
-export interface IAutoClosingPairConditional extends IAutoClosingPair {
-	notIn?: string[];
-
-}
-
-/**
- * Interface used to support electric characters
- * @internal
- */
-export interface IElectricAction {
-	// Only one of the following properties should be defined:
-
-	// The line will be indented at the same level of the line
-	// which contains the matching given bracket type.
-	matchOpenBracket?: string;
-
-	// The text will be appended after the electric character.
-	appendText?: string;
-
-	// The number of characters to advance the cursor, useful with appendText
-	advanceCount?: number;
-}
-
-/**
- * Describes what to do with the indentation when pressing Enter.
- */
-export enum IndentAction {
-	/**
-	 * Insert new line and copy the previous line's indentation.
-	 */
-	None = 0,
-	/**
-	 * Insert new line and indent once (relative to the previous line's indentation).
-	 */
-	Indent = 1,
-	/**
-	 * Insert two new lines:
-	 *  - the first one indented which will hold the cursor
-	 *  - the second one at the same indentation level
-	 */
-	IndentOutdent = 2,
-	/**
-	 * Insert new line and outdent once (relative to the previous line's indentation).
-	 */
-	Outdent = 3
-}
-
-/**
- * Describes what to do when pressing Enter.
- */
-export interface EnterAction {
-	/**
-	 * Describe what to do with the indentation.
-	 */
-	indentAction: IndentAction;
-	/**
-	 * Describes text to be appended after the new line and after the indentation.
-	 */
-	appendText?: string;
-	/**
-	 * Describes the number of characters to remove from the new line's indentation.
-	 */
-	removeText?: number;
-}
-
-/**
- * Interface used to support insertion of mode specific comments.
- * @internal
- */
-export interface ICommentsConfiguration {
-	lineCommentToken?: string;
-	blockCommentStartToken?: string;
-	blockCommentEndToken?: string;
-}
-
-export interface IAutoClosingPair {
-	open: string;
-	close: string;
-}
-
-/**
- * @internal
- */
-export interface IRichEditBrackets {
-	maxBracketLength: number;
-	forwardRegex: RegExp;
-	reversedRegex: RegExp;
-	brackets: editorCommon.IRichEditBracket[];
-	textIsBracket: { [text: string]: editorCommon.IRichEditBracket; };
-	textIsOpenBracket: { [text: string]: boolean; };
-}
-
 // --- feature registries ------
 
 /**
