+++ src/vs/editor/common/modes/languageConfiguration.ts
@@ -1,0 +1,183 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+'use strict';
+
+/**
+ * Describes how comments for a language work.
+ */
+export interface CommentRule {
+	/**
+	 * The line comment token, like `// this is a comment`
+	 */
+	lineComment?: string;
+	/**
+	 * The block comment character pair, like `/* block comment *&#47;`
+	 */
+	blockComment?: CharacterPair;
+}
+
+/**
+ * The language configuration interface defines the contract between extensions and
+ * various editor features, like automatic bracket insertion, automatic indentation etc.
+ */
+export interface LanguageConfiguration {
+	/**
+	 * The language's comment settings.
+	 */
+	comments?: CommentRule;
+	/**
+	 * The language's brackets.
+	 * This configuration implicitly affects pressing Enter around these brackets.
+	 */
+	brackets?: CharacterPair[];
+	/**
+	 * The language's word definition.
+	 * If the language supports Unicode identifiers (e.g. JavaScript), it is preferable
+	 * to provide a word definition that uses exclusion of known separators.
+	 * e.g.: A regex that matches anything except known separators (and dot is allowed to occur in a floating point number):
+	 *   /(-?\d*\.\d\w*)|([^\`\~\!\@\#\%\^\&\*\(\)\-\=\+\[\{\]\}\\\|\;\:\'\"\,\.\<\>\/\?\s]+)/g
+	 */
+	wordPattern?: RegExp;
+	/**
+	 * The language's indentation settings.
+	 */
+	indentationRules?: IndentationRule;
+	/**
+	 * The language's rules to be evaluated when pressing Enter.
+	 */
+	onEnterRules?: OnEnterRule[];
+	/**
+	 * The language's auto closing pairs. The 'close' character is automatically inserted with the
+	 * 'open' character is typed. If not set, the configured brackets will be used.
+	 */
+	autoClosingPairs?: IAutoClosingPairConditional[];
+	/**
+	 * The language's surrounding pairs. When the 'open' character is typed on a selection, the
+	 * selected string is surrounded by the open and close characters. If not set, the autoclosing pairs
+	 * settings will be used.
+	 */
+	surroundingPairs?: IAutoClosingPair[];
+	/**
+	 * **Deprecated** Do not use.
+	 *
+	 * @deprecated Will be replaced by a better API soon.
+	 */
+	__electricCharacterSupport?: IBracketElectricCharacterContribution;
+}
+
+/**
+ * Describes indentation rules for a language.
+ */
+export interface IndentationRule {
+	/**
+	 * If a line matches this pattern, then all the lines after it should be unindendented once (until another rule matches).
+	 */
+	decreaseIndentPattern: RegExp;
+	/**
+	 * If a line matches this pattern, then all the lines after it should be indented once (until another rule matches).
+	 */
+	increaseIndentPattern: RegExp;
+	/**
+	 * If a line matches this pattern, then **only the next line** after it should be indented once.
+	 */
+	indentNextLinePattern?: RegExp;
+	/**
+	 * If a line matches this pattern, then its indentation should not be changed and it should not be evaluated against the other rules.
+	 */
+	unIndentedLinePattern?: RegExp;
+}
+
+/**
+ * Describes a rule to be evaluated when pressing Enter.
+ */
+export interface OnEnterRule {
+	/**
+	 * This rule will only execute if the text before the cursor matches this regular expression.
+	 */
+	beforeText: RegExp;
+	/**
+	 * This rule will only execute if the text after the cursor matches this regular expression.
+	 */
+	afterText?: RegExp;
+	/**
+	 * The action to execute.
+	 */
+	action: EnterAction;
+}
+
+export interface IBracketElectricCharacterContribution {
+	docComment?: IDocComment;
+}
+
+/**
+ * Definition of documentation comments (e.g. Javadoc/JSdoc)
+ */
+export interface IDocComment {
+	/**
+	 * The string that starts a doc comment (e.g. '/**')
+	 */
+	open: string;
+	/**
+	 * The string that appears on the last line and closes the doc comment (e.g. ' * /').
+	 */
+	close: string;
+}
+
+/**
+ * A tuple of two characters, like a pair of
+ * opening and closing brackets.
+ */
+export type CharacterPair = [string, string];
+
+export interface IAutoClosingPair {
+	open: string;
+	close: string;
+}
+
+export interface IAutoClosingPairConditional extends IAutoClosingPair {
+	notIn?: string[];
+}
+
+/**
+ * Describes what to do with the indentation when pressing Enter.
+ */
+export enum IndentAction {
+	/**
+	 * Insert new line and copy the previous line's indentation.
+	 */
+	None = 0,
+	/**
+	 * Insert new line and indent once (relative to the previous line's indentation).
+	 */
+	Indent = 1,
+	/**
+	 * Insert two new lines:
+	 *  - the first one indented which will hold the cursor
+	 *  - the second one at the same indentation level
+	 */
+	IndentOutdent = 2,
+	/**
+	 * Insert new line and outdent once (relative to the previous line's indentation).
+	 */
+	Outdent = 3
+}
+
+/**
+ * Describes what to do when pressing Enter.
+ */
+export interface EnterAction {
+	/**
+	 * Describe what to do with the indentation.
+	 */
+	indentAction: IndentAction;
+	/**
+	 * Describes text to be appended after the new line and after the indentation.
+	 */
+	appendText?: string;
+	/**
+	 * Describes the number of characters to remove from the new line's indentation.
+	 */
+	removeText?: number;
+}
