+++ src/vs/editor/test/common/testModes.ts
@@ -4,7 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 'use strict';
 
-import { LanguageConfigurationRegistry, CommentRule } from 'vs/editor/common/modes/languageConfigurationRegistry';
+import { LanguageConfigurationRegistry } from 'vs/editor/common/modes/languageConfigurationRegistry';
+import { CommentRule } from 'vs/editor/common/modes/languageConfiguration';
 import { MockMode } from 'vs/editor/test/common/mocks/mockMode';
 
 export class CommentMode extends MockMode {
